JIRA: RW-2661 Additional changes to reader code and cleanup unit tests. AUTHOR: foleyjp MERGE_REQUEST: 613 f2eefc8e83bdb0f4068e753420e488ad1c3f9d68
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@106556 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/app/table2asn/struc_cmt_reader.cpp
@@ -115,7 +115,7 @@ void CTable2AsnStructuredCommentsReader::x_LoadComments()
     }
 }
 
-void CTable2AsnStructuredCommentsReader::x_LogMessage(EDiagSev sev, const string& msg, unsigned int lineNum)
+void CTable2AsnStructuredCommentsReader::x_LogMessage(EDiagSev sev, const string& msg, unsigned long lineNum)
 {
     string newMsg = "In " + m_Filename + ": " + msg;
     CStructuredCommentsReader::x_LogMessage(sev, newMsg, lineNum);
